% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGCCombine.R
\name{AGCCombine}
\alias{AGCCombine}
\title{Inference using accelerated g-computation}
\usage{
AGCCombine(results, level = 0.95)
}
\arguments{
\item{results}{A data.frame with columns Iteration giving the iteration of
the sample parameter value, CopyID giving the copy number,and Param giving the
estimate of the desired causal parameter, and ParamName giving the name of
the parameter.}

\item{level}{The confidence level to be used for constructing confidence
intervals and performing hypothesis tests.}
}
\value{
A data.frame object with the following columns
\itemize{
  \item ParamName - The name of the parameter
  \item estimate  - The aggregated estimate of the parameter
  \item standard_error - The estimated standard error of the estimate
  \item between_var - The "between" component of the variance
  \item within_var - The "within" component of the variance
  \item var_est  - The estimated variance (square of standard_error)
  \item dof - The degrees-of-freedom of the Sattherthwaite approximation
  \item Z - Test statistic for testing if the parameter is zero
  \item p_value - The (two-sided) p-value for the test that the parameter is zero.
}
}
\description{
Combines the results of analysis on multiple simulated datasets. Associated
to each simulated dataset is an estimate of the parameters of interest.
Datasets are indexed by (i) the iteration of the bootstrap/MCMC scheme used
to generate the simulated data and (ii) an id indicating which of K
simulations the dataset corresponds to for a particular iteration.
}
